Change data type in schema for latitude/longitude
Tips and Hints: The data type for latitude and longitude is float.
Update latitue/longitude data types to be float
Update README.md with a question about what the differences are between data types

+++ db/schema.sql
@@ -8,6 +8,6 @@ CREATE TABLE users (
     country VARCHAR(255),
     photo TEXT,
     about TEXT,
-    latitude DECIMAL,
-    longitude DECIMAL
+    latitude FLOAT,
+    longitude FLOAT
 );
